+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K2CPSVQQED7M70WQ94HEZKDT"
+        "01K2EPRDBCD8GQGF7B88FPB0GS"
     )
     r"""The id of the resource"""