+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -74,7 +74,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-08-11T13:58:24.489Z")
+    created: Optional[datetime] = parse_datetime("2025-08-12T08:36:05.902Z")
     r"""The date and time the resource was created"""