+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetdatapointsop.py
@@ -600,7 +600,7 @@ class ListDatasetDatapointsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-08-11T13:58:21.237Z")
+    updated: Optional[datetime] = parse_datetime("2025-08-12T08:36:02.813Z")
     r"""The date and time the resource was last updated"""