+++ packages/orq-rc/RELEASES.md
@@ -2229,3 +2229,13 @@ Based on:
 - [python v3.11.0-rc.5] packages/orq-rc
 ### Releases
 - [PyPI v3.11.0-rc.5] https://pypi.org/project/orq-ai-sdk/3.11.0-rc.5 - packages/orq-rc
+
+## 2025-08-12 08:37:34
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.601.0 (2.680.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.11.0-rc.6] packages/orq-rc
+### Releases
+- [PyPI v3.11.0-rc.6] https://pypi.org/project/orq-ai-sdk/3.11.0-rc.6 - packages/orq-rc